Operations

Users can parametrize the decimal arithmetic by setting Decimals.Context using the Decimals.with_context function or the Decimals.@with_context macro. Operations affected by context are denoted by the badge .

Decimals.Context — Type

Context

User-selectable parametrization of the decimal arithmetics. The parameters are:

Parameter	Description

precision::Int	Maximum number of significand digits (default: 28)

rounding::RoundingMode	Rounding mode to be used when necessary (default: RoundNearest)

Emax::Int	Maximum adjusted exponent (default: 999999)

Emin::Int	Minimum adjusted exponent (default: -999999)

The parameters can be set via the (unexported) function with_context or macro @with_context:

julia> Decimals.with_context(precision=1) do
           +dec"0.1234"
       end
0.1

julia> Decimals.@with_context (precision=1,) +dec"0.1234"
0.1

Decimals.with_context — Function

with_context(f; kwargs...)

Run f with Context parametrized by kwargs.

Examples

julia> Decimals.with_context(precision=42) do
    precision(Decimal)
end
42

See also @with_context.

Decimals.@with_context — Macro

@with_context params expr

Run expr with Context parametrized by a named tuple params.

Examples

julia> Decimals.@with_context (precision=42, ) precision(Decimal)
42

See also with_context.

Arithmetic operations

Unary plus, minus

The unary plus, +x, and unary minus, -x, is equivalent to 0 + x and 0 - x, respectively.

julia> x = dec"1.1"
1.1

julia> +x
1.1

julia> -x
-1.1

Addition, subtraction

Addition and subtraction are implemented via the binary operators + and y.

julia> x = dec"1.1"
1.1

julia> y = dec"2.2"
2.2

julia> x + y
3.3

julia> x - y
-1.1

Multiplication, division

Multiplication and division are implemented via the binary operators * and /.

Division by zero throws DivisionByZeroError unless the dividend is also zero, in which case UndefinedDivisionError is thrown.

julia> x = dec"2"
2

julia> y = dec"3"
3

julia> x * y
6

julia> x / y
0.6666666666666666666666666667

julia> x / dec"0"
ERROR: DivisionByZeroError()

julia> dec"0" / dec"0"
ERROR: UndefinedDivisionError()

Absolute value

Absolute value is implemented via the Base.abs function.

julia> abs(dec"1")
1

julia> abs(dec"-1")
1

Minimum, maximum

Selecting the minimum and maximum of two Decimals is implemented via the Base.min and Base.max functions.

julia> x = dec"0.123"
0.123

julia> y = dec"4.567"
4.567

julia> min(x, y)
0.123

julia> max(x, y)
4.567

Comparison operations

The basic comparison operation is implemented via Base.cmp, which compares x and y and returns -1, 0, or +1 if x is less than, equal to, or greater than y.

julia> x = dec"0.123"
0.123

julia> y = dec"4.567"
4.567

julia> cmp(x, y)
-1

julia> cmp(y, x)
1

The binary operators ==, <, and <= are also implemented.

julia> x = dec"0.123"
0.123

julia> y = dec"4.567"
4.567

julia> x == y
false

julia> x < y
true

julia> x > y
false

julia> x <= y
true

julia> x >= y
false

Miscellaneous

Hashing

Hashing of Decimals is supported via the Base.hash function. It holds for the output hash code that x == y implies hash(x) == hash(y), and this is true even if one of the operands is not a Decimal.

julia> x = Decimal(0, 2125, -3)
2.125

julia> y = Decimal(0, 212500, -5)
2.12500

julia> hash(x) == hash(y)
true

julia> hash(x) == hash(2.125)
true

Rounding

The standard interface for rounding floating-point numbers is supported:

julia> x = dec"123.45678"
123.45678

julia> round(x) # Round to the nearest integer
123

julia> round(Int, x) # Round to the nearest integer and convert to Int
123

julia> round(x, RoundUp) # Round using a particular rounding mode
124

julia> round(x, digits=4) # Round to four decimal places
123.4568

julia> round(x, sigdigits=4) # Round to four significant digits
123.5

Our test suite covers six rounding modes that are specified by the Decimal arithmetic specification. The rounding modes are

RoundNearest (round-half-even in the specification),

RoundNearestTiesAway (round-half-up),

RoundUp (round-ceiling),

RoundDown (round-floor),

RoundFromZero (round-up),

RoundToZero (round-down).

The default mode is RoundNearest.

Normalization

Normalization of a Decimal removes all trailing zeros of the coefficient $c$ while adjusting the exponent $q$ so that the Decimal remains the same (up to the precision given by context). The operation is semantically equivalent to unary plus.

Decimals.normalize — Function

normalize(x::Decimal)

Return an equal number reduced to its simplest form with all trailing zeros in the coefficient removed.

Examples

julia> x = dec"1.2000"
1.2000

julia> x.c # Coefficient of `x`
12000

julia> y = normalize(dec"1.2000")
1.2

julia> y.c # Coefficient of `y`
12

julia> x == y
true

Exception types

Decimals.DivisionByZeroError — Type

DivisionByZeroError

Division was attempted with a denominator value of 0.

Decimals.UndefinedDivisionError — Type

UndefinedDivisionError

Division was attempted with both numerator and denominator value of 0.